# family-hearth-api/src/main/java/com/familyhearth/posts/dto/AuthorDto.java
package com.familyhearth.posts.dto;

import lombok.Data;

@Data
public class AuthorDto {
    private Long id;
    private String firstName;
    private String lastName;
    private String avatarUrl;
}